complete overhaul on the files utils: separation of concerns and shareable NamingRules
